requirements.txt
appdirs==1.4.3
bcrypt==3.2.0
blinker==1.4
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
pycparser==2.20
pyparsing==2.4.6
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.4.22
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==2.0.1
WTForms==2.3.3
